requirements.txt
vosk>=0.3.45
pyttsx3>=2.90
pyaudio>=0.2.14
pytesseract>=0.3.10
Pillow>=10.0.0
pynput>=1.7.6

